requirements.txt
anthropic==0.18.1
langchain==0.1.9
langchain-anthropic==0.1.1
langchain-community==0.0.21
chromadb==0.4.22
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.1
httpx==0.26.0
tiktoken==0.5.2
pandas==2.2.0
numpy==1.26.3
folium==0.15.1
streamlit-folium==0.15.1
